refactor: wrap ChatUI logic
